refactor: optimized
